Use a locals-like mechanism to track nesting level
This change overhauls the nesting-level tracking mechanism. Before,
nesting levels were defined by `QueryMatch`s. This worked with a
change to tree-sitter robustly but queries were quite ugly and
unintuitive.

This method is simpler. Instead of `@rainbow` for capturing brackets
and `QueryMatch`s for scopes, we instead have two scopes:

    [
        (tuple_expression)
        (type_arguments)
        ; ..
    ] @rainbow.scope

    [
        "(" ")"
        "[" "]"
        ; ..
    ] @rainbow.bracket

When `@rainbow.scope` matches, we push a new rainbow scope onto a
global stack of rainbow scopes. (I.e. not one rainbow scope stack
per layer. This lets us preserve nesting levels within injections
such as Rust macro invocations.) When `@rainbow.bracket` matches,
we check if the captured node's parent is the scope node. If so,
that node inherits the scope's highlight. This requires no changes
upstream in tree-sitter and the queries are far more intuitive.